refactor: divide footer in 3 components
- composite pattern
