+++ frontend/src/keys/injection.ts
@@ -1,0 +1,3 @@
+import { type InjectionKey, type ComputedRef } from 'vue';
+
+export const IsActive: InjectionKey<ComputedRef<boolean>> = Symbol('is-active');
